:sparkles: add config view user credit

+++ database/migrations/2024-01/2024_01_28_162130_add_credit_column_to_users_table.php
@@ -15,7 +15,7 @@ return new class extends Migration {
         Schema::table(
             'users',
             function (Blueprint $table) {
-                $table->integer('credit')->default(0);
+                $table->float('credit')->default(0);
             }
         );
     }
